+++ crates/client/db/src/storage_handler/contract_class_data.rs
@@ -35,7 +35,13 @@ impl StorageView for ContractClassDataView {
     }
 
     fn contains(&self, class_hash: &Self::KEY) -> Result<bool, DeoxysStorageError> {
-        Ok(self.get(class_hash)?.is_some())
+        let db = DeoxysBackend::expose_db();
+        let column = db.get_column(Column::ContractClassData);
+
+        match db.key_may_exist_cf(&column, class_hash.encode()) {
+            true => Ok(self.get(class_hash)?.is_some()),
+            false => Ok(false),
+        }
     }
 }