+++ crates/client/rpc/src/methods/read/estimate_fee.rs
@@ -1,6 +1,5 @@
 use blockifier::transaction::account_transaction::AccountTransaction;
 use jsonrpsee::core::RpcResult;
-use mc_genesis_data_provider::GenesisProvider;
 use mp_hashers::HasherT;
 use mp_simulations::convert_flags;
 use mp_transactions::from_broadcasted_transactions::ToAccountTransaction;
@@ -8,8 +7,6 @@ use mp_types::block::DBlockT;
 use pallet_starknet_runtime_api::{ConvertTransactionRuntimeApi, StarknetRuntimeApi};
 use sc_client_api::backend::{Backend, StorageProvider};
 use sc_client_api::BlockBackend;
-use sc_transaction_pool::ChainApi;
-use sc_transaction_pool_api::TransactionPool;
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use starknet_core::types::{
@@ -30,20 +27,17 @@ use crate::{utils, Starknet};
 /// # Returns
 ///
 /// * `fee_estimate` - fee estimate in gwei
-pub async fn estimate_fee<A, BE, G, C, P, H>(
-    starknet: &Starknet<A, BE, G, C, P, H>,
+pub async fn estimate_fee<BE, C, H>(
+    starknet: &Starknet<BE, C, H>,
     request: Vec<BroadcastedTransaction>,
     simulation_flags: Vec<EstimateFeeFlag>,
     block_id: BlockId,
 ) -> RpcResult<Vec<FeeEstimate>>
 where
-    A: ChainApi<Block = DBlockT> + 'static,
-    P: TransactionPool<Block = DBlockT> + 'static,
     BE: Backend<DBlockT> + 'static,
     C: HeaderBackend<DBlockT> + BlockBackend<DBlockT> + StorageProvider<DBlockT, BE> + 'static,
     C: ProvideRuntimeApi<DBlockT>,
     C::Api: StarknetRuntimeApi<DBlockT> + ConvertTransactionRuntimeApi<DBlockT>,
-    G: GenesisProvider + Send + Sync + 'static,
     H: HasherT + Send + Sync + 'static,
 {
     let substrate_block_hash = starknet.substrate_block_hash_from_starknet_block(block_id).map_err(|e| {