+++ crates/client/rpc/src/methods/read/syncing.rs
@@ -1,13 +1,10 @@
 use jsonrpsee::core::RpcResult;
-use mc_genesis_data_provider::GenesisProvider;
 use mc_sync::l2::get_highest_block_hash_and_number;
 use mp_hashers::HasherT;
 use mp_types::block::DBlockT;
 use pallet_starknet_runtime_api::{ConvertTransactionRuntimeApi, StarknetRuntimeApi};
 use sc_client_api::backend::{Backend, StorageProvider};
 use sc_client_api::BlockBackend;
-use sc_transaction_pool::ChainApi;
-use sc_transaction_pool_api::TransactionPool;
 use sp_api::ProvideRuntimeApi;
 use sp_arithmetic::traits::UniqueSaturatedInto;
 use sp_blockchain::HeaderBackend;
@@ -29,15 +26,12 @@ use crate::{madara_backend_client, Starknet};
 /// This is an asynchronous function due to its reliance on `sync_service.best_seen_block()`,
 /// which potentially involves network communication and processing to determine the best block
 /// seen by the sync service.
-pub async fn syncing<A, BE, G, C, P, H>(starknet: &Starknet<A, BE, G, C, P, H>) -> RpcResult<SyncStatusType>
+pub async fn syncing<BE, C, H>(starknet: &Starknet<BE, C, H>) -> RpcResult<SyncStatusType>
 where
-    A: ChainApi<Block = DBlockT> + 'static,
-    P: TransactionPool<Block = DBlockT> + 'static,
     BE: Backend<DBlockT> + 'static,
     C: HeaderBackend<DBlockT> + BlockBackend<DBlockT> + StorageProvider<DBlockT, BE> + 'static,
     C: ProvideRuntimeApi<DBlockT>,
     C::Api: StarknetRuntimeApi<DBlockT> + ConvertTransactionRuntimeApi<DBlockT>,
-    G: GenesisProvider + Send + Sync + 'static,
     H: HasherT + Send + Sync + 'static,
 {
     // obtain best seen (highest) block number