+++ crates/node/src/chain_spec.rs
@@ -128,7 +128,7 @@ pub fn deoxys_config(sealing: SealingMode, chain_id: &str) -> Result<DevChainSpe
     ))
 }
 
-#[allow(deprecated, dead_code)]
+#[allow(deprecated)]
 fn load_genesis() -> Result<GenesisData, String> {
     log::info!("🧪 Fetching genesis block");
     let runtime = Runtime::new().unwrap();